Fix level comparison when filtering records
At least on PHP 8.2.18 with Monolog 3.7.0 the comparison of the
`\Monolog\Level` enums using `>=` evaluates to false when the  record
level is anything other than exactly the level specified rather than
comparing the integer value of the Enum.

The solution is to use `\Monolog\Level::includes()`
https://github.com/Seldaek/monolog/blob/main/src/Monolog/Level.php#L111-L117
to handle the comparison operation rather than `>=`.